Minimal wiring to run a basic command
Extracted from a sprawling exploration of the idea and additional
features.

`Host` and `HostCollection` manage host information and translating host
specifications from a hash, string, or array.

`Command` is the interface for running commands. For now, the test
command, `Uptime`, runs a command locally instead of connecting to a
remote host.

`Runner` accepts `ARGV` and:
* Finds the current `warg` directory containing config and commands.
* Loads commands
* Finds the command from `ARGV`