Uploading files and creating them from strings
`Host#create_file_from` accepts a string, a path to place the file, and
the mode for the file.

Add a test for `Host#upload`, which existed already for uploading
scripts.

Change `Host#run_script` to use `#create_file_from`.

Define proxy methods on `HostCollection` for `Host#upload` and
`Host#create_file_from`.