Setup CI with Semaphore CI
Test ruby versions going back to 2.0 because why not. Changes made for
compatibility are:
* Using `send` for certain methods that are private in older versions of
  ruby like `attr_reader` and `define_method`
* Defining `Hash#transform_keys` when not defined
* Wrapping `"$HOME"` in a `Pathname` instance because
  `Pathname#relative_path_from` expects a `Pathname` instance in older
  versions.
* Define `Host#hash` for equality comparisons.

A challenge on all CI providers surveyed is the lack of having `vagrant`
installed.

To work around this, use an Ansible playbook to create a `warg`
user on the testing VM and the necessary SSH keypairs to allow the
testing VM user to SSH into `localhost` as `warg`.

For compatibility with development, where `vagrant` is used, the
playbook copies the `vagrant` user's `authorized_keys` entries to
`warg`'s. This allows using the `vagrant`-managed keypair to SSH into
the VM as `warg`.

Tests and `tests/ssh_config` are changed to use `warg` as the user.