Logging command success / failure for each host
Logging prints out:
* Script name of the command class being run
* The command or script being run on hosts
* A status line for each host that is updated as the command begins,
  starts, or fails.

`Console` maintains a history of what's printed and uses this when
reprinting output. History entries maintain a copy of what was printed
and their row in the history. This allows determining the diff of rows
in the content and reprinting all subsequent entries.

To maintain an accurate history, replace `$stdout` and `$stderr` with
`Console::IOProxy` instances that wrap `$stdout` and `$stderr` and
define `#puts` and `#print` to use `Console#print` and `Console#puts`
respectively.

Move `HostCollection#run_tracking_result` to `Command#run_on_hosts`.
To avoid involving `HostCollection` in console printing, `Command`
needed a generic method that `Command#run_command` and `Command#run_script` would call.
Since tracking results was added to support exiting execution of
commands on failure, it made sense to implement it in `Command#execute`
(the new method `Command#run_script` and `Command#run_command` call).

For testing, redirect `Warg.console` to an instance of `StringIO` to
prevent interfering with minitest's output. Also restore `$stdout` and
`$stderr` to `STDOUT` and `STDERR` respectively. At the moment it's
unclear how to effectively test console behavior since it depends on a
TTY.

For now, `test/console-playground` steps through updating the
content for three hosts out of order from how they're initially printed
to the console. `sleep` statements are used to add pauses between
updates to make it easier to see issues with each step; othewise the
updates are so quick it's not possible to see where the error might be.