Refactor `Warg::Console` internals and API
The previous method of re-printing content assumed all content terminated
with a newline. The new implementation supports multiple pieces of
content on the same line and content spanning multiple lines that does
not terminate in a newline.

`Console#print_content` now supports arbitrary text or text objects as well
as instances of `Content` or `HostStatus`.

Improve compatibility with `IO` objects by:
* Updating `Console#puts` so it behaves like `IO#puts` (only adding a
  terminating newline if missing).
* Updating `Console#puts` and ``Console#print` to accept 0-to-many
  arguments.
* Adding `Console#write` to be compatible with `Kernel#print` and
  `Kernel#puts` via `$stdout.print` and `$stdout.puts`.

Change `Console::CursorPosition` to so that the starting row is numbered
1 and the first column is numbered 1.

Replace `Console::TextRendering` with `Console::SGR()`. Simpler to have
a slightly longer, globally-accessible method invocation than to
`include` a module in all the contexts it is used. Define `Command#SGR`
as a proxy for `Console::SGR()`.

Initializing a console no longer replaces `$stdout` and `$stderr`. They
are temporarily replaced with `Console#redirecting_stdout_and_stderr`.

Create `bin/console-playground` as the main test for `Warg::Console`. It
isn't possible to test console CSI and SGR sequences in ruby. The best
way to verify the edge cases is by verifying sample output testing those
edge-cases.