Support dynamically adding commands
Calling `chain` within a deferred action will fail because it isn't
possible to modify a hash while iterating over it. Even if it were, it
would mean that the commands are added at the end of the sequence, not
after the current command.

The `Playlist` object tracks a list of commands and two states, which
item in the list is being run, `@played_at`, and where in the list items
are being added at, `@insert_at`.

`Playlist#add` increments `@insert_at` after adding a command. When
commands are being played, `@insert_at` is reset to `@playing_at + 1` so
after each command, new commands are inserted after the current command
being run in `Playlist`.